Botania - BoP update [not ready to merge]
-Deleted/renamed/adjusted textures of Botania and BoP

P.S.: I have ready the TODO files for Botania, BoP and Bibliocraft but I
am concerned about the size of the TODO folder: it's already 4Mb, the
files I would add are 2 Mb, for a total of 6 Mb, and in the futrure will
become bigger an bigger. Is there a way to not include the TODO folder
in download of the pack so that we can add files freely?